PWM and Filtering
 Look up the spec sheet for the RGB LED and connect the R, G, and B LEDs to different PWM pins and
write a sketch to change the color of the LED to all of the 2
24 possible colors.
2. Write a sketch making your Arduino AnalogWrite a steadily increasing voltage, then a steadily
decreasing voltage, etc., repeang to essenally generate a triangle wave voltage at a frequency of
~10 Hz.
a. Plot the raw PWM signal in serial ploer
b. Construct a low pass filter with a 3dB point of 1000 Hz; Input the PWM out into this filter;
Analog read the filtered signal and plot it
c. Repeat b with a low pass filter with a 3dB point of 10 Hz
d. Repeat b with a low pass filter with a 3dB point of 1 Hz
e. Repeat this for a sine wave generated with AnalogWrite.
3. Make a distance sensor using LEDs and phototransistor
a. Relate the LED-phototransistor distance to measured signal for the red and green LEDs by
taking data and making a graph
b. Does it make a difference if you do it in the dark?
c. Is there a difference between red and green light in the sensivity of the phototransistor?
i. Make sure if you are trying to compare them that you control for the
luminosity/brightness of each LED.
d. Is there an angular orientaon dependence for the LED and phototransistor?
e. Does this work for scaered light? (i.e. have the LED and detector facing outward and in the
same direcon, and block any direct transmission of light from the LED to the detector.) Can
you detect objects that are close by through sensing of light scaering off of them? Can you
get the range of these objects from the detector?