I2C temp measurement using MCP9808
1. Read the temperature over I2C using the MCP 9808. Also measure using a thermistor and Analog
Input.
a. What is your resolution with the MCP 9808 and the thermistor?
b. How fast can you measure the temperature using both of these methods? Compare
measurements made with a 10 ms delay, 100 ms delay, and 1000 ms delay.
i. Change the chip temperature resolution to 0.5C and repeat
2. MCP 9808 only. Set the chip alarm temperatures to 3 degrees above and below room temperature.
Read the alert pin with the Arduino to know if the alarm happens. Have the Arduino ask the chip if
the alert temperature is below the lower alarm temperature or above the alarm temperature. If it is
below, make your RGB LED blue. If it is above, make it red. If it is in the non-alarm range, make the
RGB LED green. Don’t forget that the alert output needs a pullup resistor (p 30 spec sheet).